Move server-related code to server directory
